Improving CLI
This commit moves away from OptionParser to Thor. I've used Thor for a
number of projects and it's my goto CLI library. This changes the CLI
behavior. Now there's a `lookup` subcommand. I made the metadata a flag
that can be passed explicitly instead of just chomping the remainder of
ARGV.