Bug: multiple lookups with hashes sometimes caused exception
This bug can be simulated by having two lookups, where the first
lookup is valid but returns no data, so the second lookup is used.
In this scenario, an exception is raised because the first lookup
sets the answers payload to {} and the second lookup tries to
append to it.

`test/fixtures/etc/jerakia/policy.d/default.rb` has a policy file
that simulates it, if you look up a cascading hash using this policy
it will fail with an exception.

This commit changes the code flow and gathers all the responses from
all the lookups first, before parsing the answer object.