Add config method to lookup plugins.
This adds the capability to plugins to be configured directly from
jerakia.yaml.  Config can be placed in the 'plugins' hash under a key
representing the plugin name, Jerakia::Lookup::PluginConfig is a class
who's instances expose the underlying hash to the lookup plugin from
the config method.

Eg: the following configuration in jerakia.yaml

```yaml
plugins:
  myplugin:
    setting: value
```

is referenced in the lookup plugin myplugin as

```
config[:setting]  # => "value"
```

Currently the hash is also passed as a hash to the autorun method of
the pllugin - this will be deprecated in a future version.