Bugfix: Datasource option defaults not being set.
Due to a bug in the option handling of the new datasource API in
2.0.0 defaults for datasource options were not being set.  This
can be tested by the removal of `:format => :yaml` in the test
policy as this should be set by the file datasource as a default.

This affects datasources that declare options like;

`option :foo, :default => :bar`

Also this commit fixes a typo with the `extension` option of the
file resource that was mistyped `extention` and therefore meant
this option would not work.